datacamp/pr_reviewer/pyproject.toml
[project]
name = "pr-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.3",
    "notion-client>=2.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
